+++ CHANGELOG.md
@@ -58,5 +58,10 @@ The first release of LambdaBetterGrass!
  - Updated to Java 16.
  - Updated [SpruceUI].
 
+### 1.2.0
+
+ - Added a way to disable temporarily the better layer feature. Thread-bound.
+ - Fixed an issue in layer data loading, it wasn't respecting metadata namespace.
+
 [SpruceUI]: https://github.com/LambdAurora/SpruceUI
 [Connected Block Textures]: https://www.curseforge.com/minecraft/mc-mods/connected-block-textures "Connected Block Textures CurseForge page"